refactored to modules and fixed unit tests
