+++ __tests__/common/adapters/DynamodbAdapter.int.js
@@ -1,4 +1,4 @@
-const DynamoDbAdapter = require('../../../src/common/adapters/DynamoDbAdapter.js')
+const DynamoDbAdapter = require('../../../src/common/adapters/DynamoDbAdapter.mjs')
 
 describe('DynamoDB Adapter', () => {
   it('should query by field', async () => {
