+++ __tests__/createItem/businessLogic.test.mjs
@@ -1,4 +1,4 @@
-const businessLogic = require('../../src/createItem/businessLogic.js')
+import { performCalculation } from '../../src/createItem/businessLogic.mjs'
 
 describe('Creat Item buiness logic suite', () => {
   it('should add numbers', () => {
@@ -8,7 +8,7 @@ describe('Creat Item buiness logic suite', () => {
     const method = 'add'
 
     // WHEN
-    const actual = businessLogic.performCalculation({ a, b, method })
+    const actual = performCalculation({ a, b, method })
 
     // THEN
     expect(actual).toBe(7)
@@ -23,7 +23,7 @@ describe('Creat Item buiness logic suite', () => {
     // WHEN
     let error
     try {
-      businessLogic.performCalculation({ a, b, method })
+      performCalculation({ a, b, method })
     } catch (e) {
       error = e
     }
