+++ config/stage.cjs
@@ -1,8 +1,8 @@
-import { userInfo } from 'os';
+const os = require('os');
 
 const stage = () => {
   console.log(`Stage not provided. Using "local" stage name based on username: 'dev-${os.userInfo().username}'.`);
-  return `dev-${userInfo().username}`
+  return `dev-${os.userInfo().username}`
 }
 
-export const userStage = stage
+module.exports.userStage = stage
