+++ package.json
@@ -6,10 +6,10 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "eslint": "node_modules/.bin/eslint src/**/*.js --ignore-pattern node_modules/",
+    "eslint": "node_modules/.bin/eslint src/**/*.mjs --ignore-pattern node_modules/",
     "export-env-dev": "STAGE=${STAGE:=dev} && sls export-env --all -s $STAGE",
     "export-env-local": "sls export-env --all",
-    "test": "jest --config config/unit.jest.config.js",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --config config/unit.jest.config.js",
     "integration": "npm run export-env-local && jest --config config/integration.jest.config.js",
     "int": "jest --config config/integration.jest.config.js",
     "e2e": "jest --config config/e2e.jest.config.js",
@@ -19,14 +19,14 @@
   "devDependencies": {
     "@aws-sdk/client-dynamodb": "^3.465.0",
     "@aws-sdk/lib-dynamodb": "^3.465.0",
-    "@types/jest": "^28.1.3",
+    "@types/jest": "^29.5.11",
     "aws-testing-library": "^1.1.1",
     "axios": "^0.27.2",
     "dotenv": "^16.0.1",
     "eslint": "^8.55.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
-    "jest": "^28.1.1",
+    "jest": "^29.7.0",
     "serverless": "^3.38.0",
     "serverless-export-env": "^2.2.0",
     "serverless-iam-roles-per-function": "^3.2.0",
