+++ src/common/adapters/DynamoDbAdapter.mjs
